multiple folder changing, including splitting of treeView into two components, added search component, started the tab layout setup on treeView component
