+++ apps/lumina-core/src/components/tabs/treeView/treeView.tsx
@@ -1,13 +1,13 @@
 "use client";
 
-import styles from "@/components/treeView/treeView.module.scss";
+import styles from "@/components/tabs/treeView/treeView.module.scss";
 import cx from "classnames";
 import { useLuminaContext } from "@/context/contextProvider";
 import { IComponentData, IComponentProps } from "@/data/data";
 import { useCallback, useState } from "react";
-import { ShowEdit } from "../showEdit/showEdit";
+import { ShowEdit } from "../../showEdit/showEdit";
 import { configs } from "@/staticComponentsPath";
-import { Button } from "../button/buttons";
+import { SearchBar } from "../../search/search";
 
 const TreeBranch = ({ data, noUp, noDown }: { data: IComponentData, noUp: boolean, noDown: boolean }) => {
   const [showChildren, setShowChildren] = useState(false);
@@ -64,41 +64,29 @@ const ComponentTree = ({ data }: { data: IComponentData[] }) => {
   );
 };
 
-export const TreeView = () => {
+export const TreeViewTab = () => {
   const {
     state: { builderDataContext },
-    dispatch,
   } = useLuminaContext();
-
-  const handleAddPageClick = useCallback(() => {
-    dispatch({
-      type: "createPage",
-      data: { name: "testPage", friendlyName: "Test Page" },
-    });
-  }, [dispatch]);
-
   return (
     <>
-      {(Object.keys(builderDataContext.builderData).length && (
-        <div>
-          <div className={styles.treeHead}>
-            <h3>Pages</h3>
-            <Button text="Add new page" color="primary" outline onClick={handleAddPageClick} iconRight="lumina-plus" />
+    <br/>
+    <br/>
+    <br/>
+    <br/>
+    <br/>
+    <br/>
+    <br/>
+    <hr/>
+    <div className={styles.treeviewContainer}>
+    <div className={styles.treeHead}>
+          <h3 className={styles.treeTitle}>Components</h3>
+          <span className={styles.addContainer}>
+          <ShowEdit noUp={true} noDown={true} id="" data={{}} config={{ name: "page", editor: { editable: false, delete: false, children: true } }} inline={true} />
+          <h6>Add</h6>
+          </span>
           </div>
-          {Object.keys(builderDataContext.builderData).map((page) => (
-            <div className={styles.treeHead} key={page}>
-              {page}
-            </div>
-          ))}
-        </div>
-      )) ||
-        null}
-      <hr />
-      <div>
-        <div className={styles.treeHead}>
-          <h3>Components</h3>
-          <ShowEdit noUp={true} noDown={true} id="add-new-cmp" data={{}} config={{ name: "page", editor: { editable: false, delete: false, children: true } }} inline={true} />
-        </div>
+        <SearchBar/>
         <ComponentTree
           data={
             builderDataContext.builderData[builderDataContext.selectedPage].children
