--- apps/lumina-core/src/components/treeView/treeView.module.scss
@@ -1,52 +1,0 @@
-.treeContainer {
-  padding: 0 0 0 15px;
-  position: relative;
-  color: var(--lum-color-text);
-
-  &:hover {
-    background-color: silver;
-  }
-
-  .treeHead {
-    padding: 0;
-  }
-}
-
-.treeHead {
-  margin: 0 0 15px 0;
-  padding: 5px;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  gap: 20px;
-  position: relative;
-
-  h3 {
-    margin: 0;
-  }
-
-  &:last-child {
-    margin-bottom: 30px;
-  }
-}
-
-.treeHeadItem {
-  margin-bottom: 5px;
-  padding: 5px;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  gap: 20px;
-  position: relative;
-  cursor: default;
-}
-
-.btnTreeView {
-  color: var(--lum-color-text-inverse);
-  background-color: var(--lum-color-background-base);
-  border: 2px solid var(--lum-color-background-inverse);
-}
-
-.pointerTreeView {
-  cursor: pointer;
-}
